[tier2] Fixed pointer truncation on IFileReadBinary and IFileWriteBinary

+++ src/soundsystem/snd_io.cpp
@@ -20,17 +20,17 @@
 class COM_IOReadBinary : public IFileReadBinary
 {
 public:
-	int open( const char *pFileName );
-	int read( void *pOutput, int size, int file );
-	void seek( int file, int pos );
-	unsigned int tell( int file );
-	unsigned int size( int file );
-	void close( int file );
+	intp open( const char *pFileName );
+	int read( void *pOutput, int size, intp file );
+	void seek( intp file, int pos );
+	unsigned int tell( intp file );
+	unsigned int size( intp file );
+	void close( intp file );
 };
 
 
 // prepend sound/ to the filename -- all sounds are loaded from the sound/ directory
-int COM_IOReadBinary::open( const char *pFileName )
+intp COM_IOReadBinary::open( const char *pFileName )
 {
 	char namebuffer[512];
 	FileHandle_t hFile;
@@ -47,10 +47,10 @@ int COM_IOReadBinary::open( const char *pFileName )
 
 	hFile = g_pFullFileSystem->Open( namebuffer, "rb", "GAME" );
 
-	return (int)hFile;
+	return (intp)hFile;
 }
 
-int COM_IOReadBinary::read( void *pOutput, int size, int file )
+int COM_IOReadBinary::read( void *pOutput, int size, intp file )
 {
 	if ( !file )
 		return 0;
@@ -58,7 +58,7 @@ int COM_IOReadBinary::read( void *pOutput, int size, int file )
 	return g_pFullFileSystem->Read( pOutput, size, (FileHandle_t)file );
 }
 
-void COM_IOReadBinary::seek( int file, int pos )
+void COM_IOReadBinary::seek( intp file, int pos )
 {
 	if ( !file )
 		return;
@@ -66,21 +66,21 @@ void COM_IOReadBinary::seek( int file, int pos )
 	g_pFullFileSystem->Seek( (FileHandle_t)file, pos, FILESYSTEM_SEEK_HEAD );
 }
 
-unsigned int COM_IOReadBinary::tell( int file )
+unsigned int COM_IOReadBinary::tell( intp file )
 {
 	if ( !file )
 		return 0;
 	return g_pFullFileSystem->Tell( (FileHandle_t)file );
 }
 
-unsigned int COM_IOReadBinary::size( int file )
+unsigned int COM_IOReadBinary::size( intp file )
 {
 	if (!file)
 		return 0;
 	return g_pFullFileSystem->Size( (FileHandle_t)file );
 }
 
-void COM_IOReadBinary::close( int file )
+void COM_IOReadBinary::close( intp file )
 {
 	if (!file)
 		return;
